clean the files
create a seperate file for the middleware (helper functions)
